Proc macro to reduce enclave app boilerplate
Macro that simplifies the creation of an enclave app. See  enclave_apps/oak_functions_enclave_app/src/main.rs to see it in use.
