+++ enclave_apps/oak_functions_enclave_app/Cargo.toml
@@ -24,6 +24,7 @@ oak_channel = { workspace = true }
 oak_crypto = { workspace = true }
 oak_dice = { workspace = true }
 oak_restricted_kernel_sdk = { workspace = true }
+oak_restricted_kernel_sdk_proc_macro = { workspace = true }
 oak_restricted_kernel_interface = { workspace = true }
 static_assertions = "*"
 zerocopy = "*"
