+++ enclave_apps/oak_functions_enclave_app/src/main.rs
@@ -21,24 +21,13 @@
 extern crate alloc;
 
 use alloc::{boxed::Box, sync::Arc};
-use core::panic::PanicInfo;
 
-use log::info;
 use oak_core::samplestore::StaticSampleStore;
-use oak_restricted_kernel_sdk::{FileDescriptorChannel, StderrLogger};
+use oak_restricted_kernel_sdk::FileDescriptorChannel;
+use oak_restricted_kernel_sdk_proc_macro::entrypoint;
 
-static LOGGER: StderrLogger = StderrLogger {};
-
-#[no_mangle]
-fn _start() -> ! {
-    log::set_logger(&LOGGER).unwrap();
-    log::set_max_level(log::LevelFilter::Debug);
-    oak_enclave_runtime_support::init();
-    main();
-}
-
+#[entrypoint]
 fn main() -> ! {
-    info!("In main!");
     #[cfg(feature = "deny_sensitive_logging")]
     {
         // Only log warnings and errors to reduce the risk of accidentally leaking execution
@@ -63,15 +52,4 @@ fn main() -> ! {
         &mut invocation_stats,
     )
     .expect("server encountered an unrecoverable error");
-}
-
-#[alloc_error_handler]
-fn out_of_memory(layout: ::core::alloc::Layout) -> ! {
-    panic!("error allocating memory: {:#?}", layout);
-}
-
-#[panic_handler]
-fn panic(info: &PanicInfo) -> ! {
-    log::error!("PANIC: {}", info);
-    oak_restricted_kernel_interface::syscall::exit(-1);
 }
