+++ oak_restricted_kernel_sdk_proc_macro/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "oak_restricted_kernel_sdk_proc_macro"
+version = "0.1.0"
+authors = ["Juliette Pluto <julsh@google.com>"]
+edition = "2021"
+license = "Apache-2.0"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "*", features = ["full"] }
+quote = "*"
